A more complete implementation of volumes in Maya. This is a repeat of r187265, with some changes:
- Ability to import density and temperature by default
- Supports retrieving all the different grids, and doesn't create a
  fluidShape unless there is a density attribute on the geo

In the future, we might support other fields. Multi-component fields
like velocity will come out as many parts: since the Maya fluid shape
expects them to be interleaved, we might need a maya node node to
interleave them correctly, so we can't support these ones just yet.

(This is not supported unless the user has Maya 2014, since there is no
way in the 2013 or 2012 APIs to create a MFnFloatArrayData, and an
MFnDoubleArrayData connected to the inDensity, inVelocity, etc.
attributes of Maya's fluid shape create internal errors.)