Workaround a bug when saving ramp parameters
Workaround a bug when saving/loading scene files. When the first point
of the ramp is at default values (i.e. all zeroes), Maya would skip
writing the values out to file. Then, when the file is loaded back, the
first point would disappear. Workaround this bug by changing the default
position to an impossible value, so that Maya will always write out the
values.