Remove all "from __future__ import [print_function, division, with_statement, absolute_import]" statements.
Other __future__ imports (which are much less common) I left alone, though they can probably be removed too.

The only places these statements remain are in a .hip file (not worth editing just for the cleanup) and in py23convert.py (where we aren't running the import, just looking for it in other code).